+++ src/app/extension/UserSessionProvider.java
@@ -11,7 +11,7 @@ import java.util.concurrent.CompletionStage;
 public class UserSessionProvider extends play.mvc.Action.Simple {
     public CompletionStage<Result> call(Http.Context ctx) {
         String sessionIdString = ctx.session().get(CookieConstants.USER_SESSION_ID_NAME);
-        Optional<Integer> sessionId = tryParseInt(sessionIdString);
+        Optional<Long> sessionId = tryParseInt(sessionIdString);
         if(sessionId.isPresent()) {
             Optional<UserSession> userSession = UserSession.findById(sessionId.get());
             if(userSession.isPresent()) {
@@ -21,12 +21,12 @@ public class UserSessionProvider extends play.mvc.Action.Simple {
         return delegate.call(ctx);
     }
 
-    private Optional<Integer> tryParseInt(String value) {
+    private Optional<Long> tryParseInt(String value) {
         if(value == null) {
             return Optional.empty();
         }
         try {
-            int i = Integer.parseInt(value);
+            Long i = Long.parseLong(value);
             return Optional.of(i);
         } catch (NumberFormatException e) {
             return Optional.empty();
